Extract integration tests into sepearte file
